+++ scripts/remote_server/index.php
@@ -106,6 +106,15 @@ else fwrite($fh, "\n".date('Y-m-d H:i:s').' >>>>>>>>>>>>>>>>>>>> Call for action
 fwrite($fh, date('Y-m-d H:i:s').' dnsserver='.$dnsserver.", instanceserver=".$instanceserver.", allowed_hosts=".$allowed_hosts."\n");
 fwrite($fh, date('Y-m-d H:i:s').' signature='.$signature.", recalculatedsignature=".$recalculatedsignature."\n");
 
+// Compare signature and recalculatedsignature
+if ($signature != $recalculatedsignature) {
+	fwrite($fh, date('Y-m-d H:i:s')." The provided signature does not match the signature recalculated from parameters and the signature key.\n");
+
+	http_response_code(598);
+	print 'The provided signature does not match the signature recalculated from parameters and the signature key.'."\n";
+	exit();
+}
+
 if (in_array($tmparray[0], array('deploy', 'undeploy', 'deployoption', 'deployall', 'undeployall'))) {
 	if ($DEBUG) fwrite($fh, date('Y-m-d H:i:s').' ./action_deploy_undeploy.sh '.$tmparray[0].' '.$paramspace."\n");
 	else fwrite($fh, date('Y-m-d H:i:s').' ./action_deploy_undeploy.sh '.$tmparray[0].' ...'."\n");